# frontend/main/openapi.json
{
  "components": {
    "schemas": {
...
      "ApprovalOut": {
        "description": "What a sign-off produced.",
        "properties": {
          "approval_id": {
            "format": "uuid",
            "title": "Approval Id",
            "type": "string"
          },
          "approved_by": {
            "title": "Approved By",
            "type": "string"
          },
          "findings_approved": {
            "title": "Findings Approved",
            "type": "integer"
          },
          "package_revision_id": {
            "format": "uuid",
            "title": "Package Revision Id",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "approval_id",
          "package_revision_id",
          "approved_by",
          "findings_approved",
          "state"
        ],
        "title": "ApprovalOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "AI-assisted shop drawing review. The AI reads; deterministic Python decides; a reviewer signs off.",
    "title": "Graniti Vicentia V1",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/projects/{project_id}/packages/{package_id}/redline.pdf": {
      "get": {
        "description": "The signed-off redline, available only for typed findings with real stored locations.",
        "operationId": "download_redline_api_v1_projects__project_id__packages__package_id__redline_pdf_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "package_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Package Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download the signed-off evidence-grounded drawing redline",
        "tags": [
          "approvals"
        ]
      }
    },
    "/api/v1/projects/{project_id}/packages/{package_id}/report": {
      "get": {
        "description": "The signed-off workbook: the tabular audit handoff.",
        "operationId": "download_report_api_v1_projects__project_id__packages__package_id__report_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "package_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Package Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download the signed-off review as a workbook",
        "tags": [
          "approvals"
        ]
      }
    },
    "/api/v1/projects/{project_id}/packages/{package_id}/report.pdf": {
      "get": {
        "description": "The branded signed-off PDF: the readable reviewer handoff.",
        "operationId": "download_pdf_report_api_v1_projects__project_id__packages__package_id__report_pdf_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "package_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Package Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download the signed-off review as a PDF",
        "tags": [
          "approvals"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/review-sessions/{review_session_id}/approve": {
      "post": {
        "description": "Approve every finding of the revision this sitting is reviewing.\n\nThe finding set is chosen by the server, not sent by the caller. A client naming the findings it\napproves is a client that can approve a subset and leave the rest looking reviewed \u2014 and the\napproval is the record GV stands behind when a vendor disputes a dimension.\n\nRefuses while any `REVIEW_REQUIRED` finding is unaddressed. That is not a formality: an abstention\nnobody acted on is a check that did not happen, and approving around it would put a package's name\nto a question nobody answered.",
        "operationId": "approve_api_v1_projects__project_id__review_sessions__review_session_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "review_session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Review Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sign off a package, so its review can leave the building",
        "tags": [
          "approvals"
        ]
      }
    },
...
  }
}